+++ src/fillFileMap_V12.h
@@ -33,6 +33,30 @@ RA2bZinvAnalysis::fillFileMap() {
   /* fileMap_["dyee"].push_back(dir+TString("/tree_DYe_CleanVars/tree_DYJetsToLL_M-50_HT-1200to2500.root")); */
   fileMap_["dyee"].push_back(dir+TString("/tree_DYe_CleanVars/tree_DYJetsToLL_M-50_HT-2500toInf.root"));
 
+  // ttz->mumu MC
+  fileMap_["ttzmm"].push_back(dir+TString("/tree_DYm_CleanVars/tree_TTZToLLNuNu.root"));
+
+  // ttz->ee MC
+  fileMap_["ttzee"].push_back(dir+TString("/tree_DYe_CleanVars/tree_TTZToLLNuNu.root"));
+
+  // VV->mumu MC
+  fileMap_["VVmm"].push_back(dir+TString("/tree_DYm_CleanVars/tree_ZZTo2L2Q.root"));
+  fileMap_["VVmm"].push_back(dir+TString("/tree_DYm_CleanVars/tree_WWTo2L2Nu.root"));
+
+  // VV->ee MC
+  fileMap_["VVee"].push_back(dir+TString("/tree_DYe_CleanVars/tree_ZZTo2L2Q.root"));
+  fileMap_["VVee"].push_back(dir+TString("/tree_DYe_CleanVars/tree_WWTo2L2Nu.root"));
+
+  // ttbar->mumu MC
+  fileMap_["ttmm"].push_back(dir+TString("/tree_DYm_CleanVars/tree_TTJets_DiLept.root"));
+  fileMap_["ttmm"].push_back(dir+TString("/tree_DYm_CleanVars/tree_TTJets_SingleLeptFromTbar.root"));
+  fileMap_["ttmm"].push_back(dir+TString("/tree_DYm_CleanVars/tree_TTJets_SingleLeptFromT.root"));
+
+  // ttbar->ee MC
+  fileMap_["ttee"].push_back(dir+TString("/tree_DYe_CleanVars/tree_TTJets_DiLept.root"));
+  fileMap_["ttee"].push_back(dir+TString("/tree_DYe_CleanVars/tree_TTJets_SingleLeptFromTbar.root"));
+  fileMap_["ttee"].push_back(dir+TString("/tree_DYe_CleanVars/tree_TTJets_SingleLeptFromT.root"));
+
   // DY->mumu data
   fileMap_["zmm"].push_back(dir+TString("/tree_DYm_CleanVars/tree_SingleMuon_re2016B.root"));
   fileMap_["zmm"].push_back(dir+TString("/tree_DYm_CleanVars/tree_SingleMuon_re2016C.root"));
